signOut/requirements.txt
anyio==3.6.2
autopep8==2.0.1
click==8.1.3
fastapi==0.91.0
h11==0.14.0
httptools==0.5.0
idna==3.4
psycopg2==2.9.5
pycodestyle==2.10.0
pydantic==1.10.4
python-dotenv==0.21.1
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.24.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
